Anthill Iq Chatbot/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
openai>=1.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
httpx>=0.24.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # For PostgreSQL
alembic>=1.12.0  # For database migrations 